Finish moving and renaming, add friendbot export
